% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscNoise.R
\name{tune_hyper_parameters}
\alias{tune_hyper_parameters}
\title{Hyper parameter tuning function
#'}
\usage{
tune_hyper_parameters(classifier, train, n_cores)
}
\arguments{
\item{classifier}{a string, takes the name of the classifier.Currently supported classifiers are 
'rf' - Random forest
'lrm' - Logistic regression
'CART' - Classification tree
'knn' - K-Nearest Neighbors}

\item{train}{data frame to be used for hyper parameter tuning}

\item{n_cores}{No of cores to be used for the process of hyper parameter optimization}
}
\description{
Hyper parameter tuning function
#'
}
